requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database Drivers
PyMySQL==1.1.0
redis==5.0.1
opensearch-py==2.4.2

# AI/ML Libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-google-genai==0.0.5
openai==1.6.1
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# AutoGen
pyautogen==0.2.0

# Monitoring & Tracing
langfuse==2.7.3
arize-phoenix==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Utilities
certifi==2023.11.17
urllib3==2.1.0
cryptography==41.0.8